Fixed problem with confirm files.
Files were being written in one format and read in another, which caused an exception
which stopped the server from working.  The format and code was cleaned up to be consistent.
dAlso added fields in the files for debugging.  Also changed the file name format to be
simpler and more predictable.