Massive update to improve reliability.
When DICOM network operations fail, they will now reset the socket and try again.  The types of failures handled are:

   1: Operation times out.  Timeout is configuration value DicomTimeout_sec
   2: Operation returns a bad status.

Confirms are stagger-started so if there are lots of them they don't hammer the synchronized DICOM connections.

Series now use SeriesDate+SeriesTime, which is the same for each file in the series.  As opposed to AcquisitionTime and AcquisitionDate, which was previously being used.

Sent series are persisted.  Previously they were not, which meant that on startup that if there were a lot of them that they could cause a significant slowdown on startup.

Added configuration file for getting files from Varian VMSDBD.