Restart server if there is a problem closing a DICOM socket.
Sometimes the close operation throws an exception, and the most robust way to handle it is to restart.

These errors do not happen often, on the order of twice a day, so this will not significantly impact the clinic.  The restarts take around 2 minutes, which is inconvenient but not catastrophic.